JIT: Reduce jit_entry_calls/jit_exception_calls to 32 bits
We don't need 64 bits for jit_entry_calls/jit_exception_calls so dropping
them down to 32 bits will allow us to save 8 bytes per iseq.

+++ zjit.h
@@ -109,8 +109,8 @@ ZJIT_STACK_MAP_BASE_PTR_STACK_SIZE(VALUE entry)
 extern void *rb_zjit_entry;
 extern bool rb_zjit_compiling_p;
 extern const zjit_jit_frame_t rb_zjit_c_frame;
-extern uint64_t rb_zjit_call_threshold;
-extern uint64_t rb_zjit_profile_threshold;
+extern unsigned int rb_zjit_call_threshold;
+extern unsigned int rb_zjit_profile_threshold;
 void rb_zjit_compile_iseq(const rb_iseq_t *iseq, rb_execution_context_t *ec, bool jit_exception);
 void rb_zjit_profile_insn(uint32_t insn, rb_execution_context_t *ec);
 void rb_zjit_profile_enable(const rb_iseq_t *iseq);
